requirements.txt
groq
google-generativeai
cohere
streamlit
python-dotenv
requests
openai